requirements.txt
# Core Dependencies
PyPDF2==3.0.1
pdfplumber==0.10.0
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2
python-dateutil==2.8.2
tabulate==0.9.0
matplotlib==3.8.2
seaborn==0.13.0
jinja2==3.1.2
reportlab==4.0.7
xlsxwriter==3.1.9

# AI/LLM Integration
openai>=1.0.0

# File Monitoring
watchdog>=3.0.0

# Environment Management
python-dotenv>=1.0.0

# Additional Utilities
xlrd>=2.0.0
pytz>=2023.3
requests>=2.31.0